Refactored code to seperate functions used into sms router into seperate file.
